# Simple optimization to speed up LCM