#### Brief description of what is fixed or changed
Changed computation of `lcm(a,b)` from `a * b // gcd(a,b)` to `a // gcd(a,b) * b`, for LCM computation of integers `Z` and rational numbers `Q`.

**Performance Difference**:
The speed-up is very subtle for ordinary usage, but the difference becomes more evident when the numbers involved become very large.

**Tests**:
I ran 3 tests locally for lcm of pair of integers:
Test 1: `10**5` random pairs of `2048-bit` numbers, the former finished in `40s`, and the latter in `38s`.
Test 2: `10**4` random pairs of `8192-bit` numbers, the former finished in `32s`, and the latter in `31s`.
Test 3: `720**99911` and `480**12342`, the former finished in `0.6s`, and the latter in `0.55s`.

@normalhuman @jksuom Please review.