@sidhantnagpal 2018-02-24
> I tried a different version of your example to see the difference better

This looks good. Thanks.